refactor: split selector components into their own files
